Refactor conflict dialog
