+++ src/app/shared/issue/conflict-dialog/conflict-dialog.component.css
@@ -4,22 +4,7 @@
   max-height: 55vh;
 }
 
-.column.left {
-  min-width: 60vh;
-  width: 100%;
-}
-
 .full-width {
   min-width: 80vh;
   width: 100%;
-}
-
-.column.right {
-  max-width: 15%;
-  min-width: 20vh;
-  width: 100%;
-}
-
-.column {
-  padding: 10px;
 }
