+++ src/app/shared/issue/description/description.component.html
@@ -29,7 +29,7 @@
       </button>
       <button class="editor-action" *ngIf="conflict" type="button" [disabled]="isSavePending"
               mat-raised-button color="primary" (click)="viewChanges()">
-        View Changes
+        View Updated Description
       </button>
     </div>
   </div>
