+++ src/app/shared/issue/team-response/team-response.component.html
@@ -33,7 +33,7 @@
       </button>
       <button class="editor-action" *ngIf="conflict" type="button" [disabled]="isSavePending"
               mat-raised-button color="primary" (click)="viewChanges()">
-        View Changes
+        View Updated Response
       </button>
     </div>
   </div>
