+++ src/app/shared/view-issue/new-team-response/conflict-dialog/conflict-dialog.component.html
@@ -1,0 +1,68 @@
+<div style="display: flex; margin-bottom: 20px; align-items: center">
+  <h1 mat-dialog-title style="margin: 0">{{ 'A new response was submitted by another user'}}</h1>
+  <button mat-icon-button color="default" style="margin: 0 0 0 auto" (click)="close()">
+    <mat-icon>close</mat-icon>
+  </button>
+</div>
+
+<div mat-dialog-content style="display: flex;">
+  <div class="column left">
+    <mat-tab-group class="mat-elevation-z1" animationDuration="0ms" (selectedTabChange)="handleTabChange($event)">
+      <mat-tab label="Updated Markdown Text">
+        <div class="tab-content" *ngIf="isReady">
+          <div [innerHTML]="updatedHtml"></div>
+        </div>
+      </mat-tab>
+      <mat-tab label="Preview">
+        <div class="tab-content">
+          <markdown>{{data.teamResponse}}</markdown>
+        </div>
+      </mat-tab>
+    </mat-tab-group>
+  </div>
+
+  <div class="column right">
+
+    <div *ngIf="data.type">
+      <mat-divider></mat-divider>
+      <span class="mat-title"> Type </span>
+      <p [ngStyle]="labelService.setLabelStyle(labelService.getColorOfLabel(data.type))">
+        {{data.type || '-'}}
+      </p>
+    </div>
+
+    <div *ngIf="data.severity">
+      <mat-divider></mat-divider>
+      <span class="mat-title"> Severity </span>
+      <p [ngStyle]="labelService.setLabelStyle(labelService.getColorOfLabel(data.severity))">
+        {{data.severity || '-'}}
+      </p>
+    </div>
+
+    <div *ngIf="data.responseTag">
+      <mat-divider></mat-divider>
+      <span class="mat-title"> Response </span>
+      <p [ngStyle]="labelService.setLabelStyle(labelService.getColorOfLabel(data.responseTag))">
+        {{data.responseTag || '-'}}
+      </p>
+    </div>
+
+    <div *ngIf="data.assignees">
+      <mat-divider></mat-divider>
+      <app-assignee-component  [issue]="data" [team]="data.teamAssigned"
+                               [isEditable]="false">
+      </app-assignee-component>
+    </div>
+
+    <div *ngIf="data.duplicateOf">
+      <mat-divider></mat-divider>
+      <div>
+        <span class="mat-title">A Duplicate Of:</span>
+        <p style="margin-top: 5px">
+          {{ ('#' + data.duplicateOf + ': ' + (issueService.getIssue(data.duplicateOf) | async).title) }}
+        </p>
+      </div>
+    </div>
+
+  </div>
+</div>
