+++ src/app/shared/view-issue/new-team-response/new-team-response.component.css
